// src/app/students/models/Student.ts
export interface Student {
  code: number;
  idType: string;
  idNumber: number;
  names: string;
  lastNames: string;
  collegeEmail: string;
  alternativeEmail: string;
  city: string;
  workCompany: string;
  undergraduateDegree: string;
  phoneNumber: number;
  gender: string;
  typePopulation: string;
  ethnicity: string;
  disability: string;
  masterStatus: string;
  modality: string;
  academicSemester: number;
  financialSemester: number;
}